fixed an issue where rooms were being overwritten and deleted.
In the following order, the room created by clientC was overwritten and deleted, and the rooms of clientC and clientD were not created.
To solve this problem, we fixed it so that it does not do anything if
the room does not exist when exiting, and when the other's leave is received, the room deletion is not performed.

- clientA and clientB are in the room
- clientA left (room deleted)
- clientC join (room created)
- clientB left (room deleted)
- clientD join (room created)